requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
packaging==23.2
pymongo==4.6.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
